Update PieSS_2023-01-03_ChatGPT_01.py
This program uses the `iss` and `geolocation` modules, which you will need to import. The `iss` module is a Python library for retrieving data about the International Space Station, and the `geolocation` module is a simple utility for getting the GPS location of the Raspberry Pi.

To use this program, you will need to install the `pigpio` library and connect your servo and LEDs to the Raspberry Pi according to the pin numbers specified in the constants at the top of the program.

I hope this helps! Let me know if you have any questions or if you need further assistance.