Add files via upload
Prompt:
Please generate a python program for raspberry pi 3 b+.  I'd like the program to check every 5 seconds and notify me when the ISS will be visible in the night sky when it passes over my location based on the GPS location of the raspberry pi, or by inputting my longitude and latitude manually. 10 minutes before the ISS will be visible at the defined location, i'd like an LED to be turned on. When the ISS will be visible in 5 minutes, I'd like the first LED to turn off and a separate LED to turn on to notify me that the ISS is 5 minutes away from being visible. Then, turn off the 5 minute LED and turn on another LED when the ISS is 1 minute away from being visible at the defined location. When the ISS is 5 seconds from being visible at the defined location, turn off the notification LEDs and move a servo motor from 0 degrees to 90 degrees. use the pigpio library to manage the servo. Duty time for the servo at 90 degrees is 530. Duty time for the servo at 0 degrees is 1530. Please also include a function in the program to notify me of which direction to look in the sky when the ISS is going to pass over by  illuminating a separate North, East, South, or west LED light.